requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
fastapi==0.115.6
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jax==0.4.38
jaxlib==0.4.38
jiter==0.8.2
kiwisolver==1.4.8
matplotlib==3.10.0
mediapipe==0.10.20
ml_dtypes==0.5.0
multidict==6.1.0
numpy==1.26.4
openai==1.57.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
pip==24.0
propcache==0.2.1
protobuf==4.25.5
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scipy==1.15.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
starlette==0.41.3
tqdm==4.67.1
twilio==9.4.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
websockets==14.1
yarl==1.18.3
